requirements.txt
selenium
webdriver-manager
python-dotenv
pyautogui
